Fix build by not minifying
The error reported in Issue #94 is caused by source code being minified:
https://docs.angularjs.org/error/$injector/unpr?p0=eProvider%20%3C-%20e
I don't understand why exactly it happens. We are using
ng-annotate-webpack-plugin to handle that, but for some reason it
stopped working.

Without minification everything worked great, so I tried how big would
the electron app be without minification and it turns out that at least
the Mac version is almost the same (~40MB vs. ~41MB), so I decided to
remove minification.

It has a very pleasant side effect: build time goes from ~3 minutes
to ~11 seconds.

The app.js goes from ~2MB to ~9MB, so if we want an official web version
in the future, we have to revisit this and optimize the webpack build.
But right now it is ok for the Electron app.