If a user is already a delegate, don't display the form
Solves https://github.com/LiskHQ/lisk-nano/issues/946.

Initially I attempted a redirect to the root page here so the modal
wouldn’t load if the user is a delegate. Due to the way this dialog is
opened though, that caused the dialog to quickly animate onto the page,
then disappear after the redirect was triggered. This is a quick
compromise for the UX.